--- app/src/main/java/com/example/dikidimainscreen/data/AuthorizationAPI.kt
@@ -1,12 +1,0 @@
-package com.example.dikidimainscreen.data
-
-import com.example.dikidimainscreen.domain.model.DikidiServerResponse
-import io.reactivex.Single
-import retrofit2.http.GET
-
-interface AuthorizationAPI {
-
-@GET("/home/info")
-//fun getExampleData(@Header("Authorization") token: String): Call<DikidiServerResponse>
-fun getExampleData(): Single<DikidiServerResponse>
-}
