+++ app/src/main/java/com/example/dikidimainscreen/data/model/Catalog.kt
@@ -1,0 +1,14 @@
+package com.example.dikidimainscreen.data.model
+
+
+import com.google.gson.annotations.SerializedName
+
+data class Catalog(
+    @SerializedName("id")
+    val id: Int,
+    @SerializedName("image")
+    val image: Image,
+    @SerializedName("name")
+    val name: String,
+
+)
