+++ app/src/main/java/com/example/dikidimainscreen/domain/model/CatalogData.kt
@@ -1,6 +1,7 @@
-package com.example.dikidimainscreen.domain.testdata
+package com.example.dikidimainscreen.domain.model
 
 data class CatalogData (
+    val id: Int,
     val imageId: Int,
     val title: String
 )
