--- app/src/main/java/com/example/dikidimainscreen/domain/model/Currency.kt
@@ -1,13 +1,0 @@
-package com.example.dikidimainscreen.domain.model
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Currency(
-    @SerializedName("abbr")
-    val abbr: String,
-    @SerializedName("id")
-    val id: String,
-    @SerializedName("title")
-    val title: String
-)
