--- app/src/main/java/com/example/dikidimainscreen/domain/model/Schedule.kt
@@ -1,13 +1,0 @@
-package com.example.dikidimainscreen.domain.model
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Schedule(
-    @SerializedName("day")
-    val day: String,
-    @SerializedName("work_from")
-    val workFrom: String,
-    @SerializedName("work_to")
-    val workTo: String
-)
