--- app/src/main/java/com/example/dikidimainscreen/domain/model/Shares.kt
@@ -1,11 +1,0 @@
-package com.example.dikidimainscreen.domain.model
-
-
-import com.google.gson.annotations.SerializedName
-
-data class Shares(
-    @SerializedName("count")
-    val count: String,
-    @SerializedName("list")
-    val list: List<SharesData>
-)
